Feature/disruptor (#7)
* oiuoiu

* asd

* iuy

* iuiuy

* 97898

* fix

* фыв

* kjhkjh

* jkhkjh

* фыв

* qwqwe

* qwe

* asd

* qwe

* dfsdf

* iuoiu

* yiuyi

* asfdf

* iuyi

* qwe

* 465

* asd

* asd

* asd

* wer

* qwe

* asdasd

* asd

* ewr

* asdasd

* asdasd

* wer

* dsasd

* uy

* asd

* oiuoi

* kjh

* qweq

* poi

+++ src/Vertr.Exchange.Common/Binary/Commands/BatchAddAccountsCommand.cs
@@ -4,9 +4,9 @@ namespace Vertr.Exchange.Common.Binary.Commands;
 
 public sealed class BatchAddAccountsCommand : IBinaryCommand
 {
-    // ExtKey = uud
-    // IntKey = symbol
-    // Value = balance
-    public IDictionary<int, IDictionary<int, long>> Users { get; set; } =
-        new Dictionary<int, IDictionary<int, long>>();
+    // LongKey = uud
+    // IntKey = currency
+    // DecimalValue = balance
+    public IDictionary<long, IDictionary<int, decimal>> Users { get; set; } =
+        new Dictionary<long, IDictionary<int, decimal>>();
 }
